+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/database/passwordoverride/JiraUserPasswordOverridingDatabase.kt
@@ -9,13 +9,13 @@ import org.apache.logging.log4j.Logger
 import java.net.URI
 import java.util.function.Function
 
-class JiraUserPasswordOverridingDatabase internal constructor(
+class JiraUserPasswordOverridingDatabase private constructor(
     private val databaseDelegate: Database,
-    private val sqlClient: SshSqlClient,
     private val username: String,
-    private val jiraDatabaseSchemaName: String,
-    private val userPasswordPlainText: String,
-    private val userPasswordEncryptorProvider: JiraUserPasswordEncryptorProvider
+    private val plainTextPassword: String,
+    private val passwordEncryption: Function<String, String>,
+    private val sqlClient: SshSqlClient,
+    private val schema: String
 ) : Database {
     private val logger: Logger = LogManager.getLogger(this::class.java)
 
@@ -26,61 +26,57 @@ class JiraUserPasswordOverridingDatabase internal constructor(
         ssh: SshConnection
     ) {
         databaseDelegate.start(jira, ssh)
-        val userPasswordEncryptor = userPasswordEncryptorProvider.getEncryptor(ssh, sqlClient)
-        val password = userPasswordEncryptor.getEncryptedPassword(userPasswordPlainText)
-        sqlClient.runSql(ssh, "UPDATE ${jiraDatabaseSchemaName}.cwd_user SET credential='$password' WHERE user_name='$username';")
-        logger.debug("Password for user '$username' updated to '${userPasswordPlainText}'")
+        val methodSelect = "SELECT attribute_value FROM $schema.cwd_directory_attribute" +
+            " WHERE attribute_name = 'user_encryption_method';"
+        val encryptionMethod = sqlClient.runSql(ssh, methodSelect).output
+        val password = when {
+            encryptionMethod.contains("plaintext") -> plainTextPassword
+            encryptionMethod.contains("atlassian-security") -> passwordEncryption.apply(plainTextPassword)
+            else -> throw RuntimeException("Unknown jira user password encryption type")
+        }
+        val passwordUpdate = "UPDATE $schema.cwd_user SET credential='$password'" +
+            " WHERE user_name='$username';"
+        sqlClient.runSql(ssh, passwordUpdate)
+        logger.debug("Password for user '$username' updated to '$plainTextPassword'")
     }
-
 
     class Builder(
         private var databaseDelegate: Database,
-        private var userPasswordPlainText: String,
-        private var userPasswordEncryptorProvider: JiraUserPasswordEncryptorProvider
+        private var passwordEncryption: Function<String, String>
     ) {
         private var sqlClient: SshSqlClient = SshMysqlClient()
-        private var jiraDatabaseSchemaName: String = "jiradb"
+        private var schema: String = "jiradb"
         private var username: String = "admin"
+        private var plainTextPassword: String = "admin"
 
         fun databaseDelegate(databaseDelegate: Database) = apply { this.databaseDelegate = databaseDelegate }
         fun username(username: String) = apply { this.username = username }
-        fun userPasswordPlainText(userPassword: String) = apply { this.userPasswordPlainText = userPassword }
+        fun plainTextPassword(passwordPlainText: String) = apply { this.plainTextPassword = passwordPlainText }
         fun sqlClient(sqlClient: SshSqlClient) = apply { this.sqlClient = sqlClient }
-        fun jiraDatabaseSchemaName(jiraDatabaseSchemaName: String) = apply { this.jiraDatabaseSchemaName = jiraDatabaseSchemaName }
-        fun userPasswordEncryptorProvider(userPasswordEncryptorProvider: JiraUserPasswordEncryptorProvider) =
-            apply { this.userPasswordEncryptorProvider = userPasswordEncryptorProvider }
+        fun schema(jiraDatabaseSchemaName: String) = apply { this.schema = jiraDatabaseSchemaName }
+        fun passwordEncryption(passwordEncryption: Function<String, String>) =
+            apply { this.passwordEncryption = passwordEncryption }
 
         fun build() = JiraUserPasswordOverridingDatabase(
             databaseDelegate = databaseDelegate,
             sqlClient = sqlClient,
             username = username,
-            userPasswordPlainText = userPasswordPlainText,
-            jiraDatabaseSchemaName = jiraDatabaseSchemaName,
-            userPasswordEncryptorProvider = userPasswordEncryptorProvider
+            plainTextPassword = plainTextPassword,
+            passwordEncryption = passwordEncryption,
+            schema = schema
         )
     }
 
 }
 
 /**
- * @param passwordEncryptFunction Based on [retrieving-the-jira-administrator](https://confluence.atlassian.com/jira/retrieving-the-jira-administrator-192836.html)
+ * @param passwordEncryption Based on [retrieving-the-jira-administrator](https://confluence.atlassian.com/jira/retrieving-the-jira-administrator-192836.html)
  * to encode the password in Jira format use [com.atlassian.crowd.password.encoder.AtlassianSecurityPasswordEncoder](https://docs.atlassian.com/atlassian-crowd/4.2.2/com/atlassian/crowd/password/encoder/AtlassianSecurityPasswordEncoder.html)
  * from the [com.atlassian.crowd.crowd-password-encoders](https://mvnrepository.com/artifact/com.atlassian.crowd/crowd-password-encoders/4.2.2).
- *
  */
-fun Database.withAdminPassword(adminPasswordPlainText: String, passwordEncryptFunction: Function<String, String>): Database {
-    val jiraDatabaseSchemaName = "jiradb"
-    val sqlClient = SshMysqlClient()
-    return JiraUserPasswordOverridingDatabase.Builder(
-        databaseDelegate = this,
-        userPasswordPlainText = adminPasswordPlainText,
-        userPasswordEncryptorProvider = DefaultJiraUserPasswordEncryptorProvider(
-            jiraDatabaseSchemaName = jiraDatabaseSchemaName,
-            plainTextPasswordEncryptor = PlainTextJiraUserPasswordEncryptor(),
-            encryptedPasswordEncryptor = EncryptedJiraUserPasswordEncryptor(passwordEncryptFunction)
-        )
-    )
-        .jiraDatabaseSchemaName(jiraDatabaseSchemaName)
-        .sqlClient(sqlClient)
-        .build()
-}
+fun Database.overridePassword(
+    passwordEncryption: Function<String, String>
+): JiraUserPasswordOverridingDatabase.Builder = JiraUserPasswordOverridingDatabase.Builder(
+    databaseDelegate = this,
+    passwordEncryption = passwordEncryption
+)