+++ src/test/kotlin/com/atlassian/performance/tools/infrastructure/api/database/passwordoverride/JiraUserPasswordOverridingDatabaseTest.kt
@@ -1,7 +1,6 @@
 package com.atlassian.performance.tools.infrastructure.api.database.passwordoverride
 
 import com.atlassian.performance.tools.infrastructure.api.database.Database
-import com.atlassian.performance.tools.infrastructure.database.SshSqlClient
 import com.atlassian.performance.tools.infrastructure.mock.MockSshSqlClient
 import com.atlassian.performance.tools.infrastructure.mock.RememberingDatabase
 import com.atlassian.performance.tools.infrastructure.mock.RememberingSshConnection
@@ -10,6 +9,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.Before
 import org.junit.Test
 import java.net.URI
+import java.util.function.Function
 
 class JiraUserPasswordOverridingDatabaseTest {
 
@@ -27,21 +27,18 @@ class JiraUserPasswordOverridingDatabaseTest {
         underlyingDatabase = RememberingDatabase()
         sshConnection = RememberingSshConnection()
         sqlClient = MockSshSqlClient()
-        database = JiraUserPasswordOverridingDatabase
-            .Builder(
-                databaseDelegate = underlyingDatabase,
-                userPasswordPlainText = samplePassword,
-                userPasswordEncryptorProvider = object : JiraUserPasswordEncryptorProvider {
-                    override fun getEncryptor(ssh: SshConnection, sqlClient: SshSqlClient): JiraUserPasswordEncryptor {
-                        return object : JiraUserPasswordEncryptor {
-                            override fun getEncryptedPassword(plainTextPassword: String) = expectedEncryptedPassword
-                        }
-                    }
-                }
-            )
+        database = underlyingDatabase.overridePassword(Function { expectedEncryptedPassword })
+            .plainTextPassword(samplePassword)
             .sqlClient(sqlClient)
-            .jiraDatabaseSchemaName("jira")
+            .schema("jira")
             .build()
+        sqlClient.queueReturnedSqlCommandResult(
+            SshConnection.SshResult(
+                exitStatus = 0,
+                output = "atlassian-security",
+                errorOutput = ""
+            )
+        )
     }
 
     @Test
@@ -74,7 +71,7 @@ class JiraUserPasswordOverridingDatabaseTest {
         // then
         assertThat(sqlClient.getLog())
             .`as`("sql queries executed")
-            .containsExactly(
+            .contains(
                 "UPDATE jira.cwd_user SET credential='${expectedEncryptedPassword}' WHERE user_name='admin';"
             )
     }